Extract setup prompt rendering helpers
Move the setup header and step rendering helpers into a dedicated
`src/setup/prompts/render.rs` sibling module so
`src/setup/prompts.rs` stays below the repository's 400-line cap.

Keep the public `print_header` and `print_step` call sites unchanged by
forwarding through `prompts.rs`, which preserves the existing setup and
test imports while isolating the display-only code.
